refactor: функции для работы с датой вынесены в отдельный модуль dateUtils
